refactor(LottoReader): LottoReader 패키지 변경(domain -> application)
