+++ src/test/java/study/lotto/step2/application/LottoReaderTest.java
@@ -1,10 +1,14 @@
-package study.lotto.step2.domain;
+package study.lotto.step2.application;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import study.lotto.step2.application.LottoReader;
+import study.lotto.step2.domain.Lotto;
+import study.lotto.step2.domain.LottoResult;
+import study.lotto.step2.domain.LottoResults;
 
 import java.util.List;
 import java.util.stream.Collectors;
